llm/rag/ell-rag/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "ell-rag-simple"
version = "0.0.0"
description = "ell rag simple"
readme = "README.md"
authors = [
    { name = "Christian D'Andrea", email = "crdandre@gmail.com" },
]
dependencies = [
    "ell-ai ~= 0.0.6",
    "scikit-learn ~= 1.5.2",
]

[project.optional-dependencies]
dev = [
    "black ~= 24.4",
    "bumpver ~= 2023.1129",
    "flake8 ~= 7.1",
    "isort ~= 5.12",
    "pre-commit ~= 3.5",
]

[tool.bumpver]
current_version = "0.0.0"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message = "bump version {old_version} -> {new_version}"
commit = true
tag = true
push = true

[tool.bumpver.file_patterns]
"pyproject.toml" = [
    'current_version = "{version}"',
    'version = "{version}"'
]
